Blockouts started
Created 4 blockouts all with different ways to be played. Will talk to the team about these blockouts and what way we want the level design to go along with getting the next part ready so that levels 2 and 3 can be started.

Pushing these so that some level design can be uploaded.