BP_CartMovement Update - Steering
Steering is working now.
It is continuous movement to one side or the other while player is standing on the lever.
Next update: maybe we could do snapping instead?
Future update: Once level is completed, we need to sort out the numbers so the cart can't go out of tracks.