Enemy Bug Update
BP_BeetleowlBug - updating in progress; fiddling with Behaviour trees and Blueprints

BP_Postman - added Player_PointLight so there is better visibility in-game. Note that the values of Light are extremely high and might need to be altered depending on the Post-processing changes. For now it is 300k Lm. Radii of different light features are also high - might need to be altered later.

Level1_LevelConstruction1_MiaTest - Lighting was built in-level and might need to be rebuilt in actual level later (Select Level > Toolbar Build>Build Lighting Only)