Postman's Parcel Carry Check in Enemy works!
BP_Postman > IsCarrying bool set to public.
In BP_EnemyBasicChar I made a check to see if player is carrying a parcel. This allows for any child enemy to see it as well.

Fiddled a lot, but in the end what worked is the addition of BTT_CheckParcelCarry and THEN adding a decorator BTD_PostmanCarryingParcel.

The Task is connected to the Interface Carry Package which is realised as Event Carry Package in the BP_EnemyBasicChar Blueprint! There was also an addition of "OnCarryEnd" Event Dispatcher in the same blueprint in order to make everything work.

Hallelu!