Healthbars working
There were certain settings that needed to be altered in Progress Bar component of the Widget. Namely, the Fill Color and Opacity should NOT have a Bind. In the blueprint settings, the DamagableActor variable should be of type BPI_Damagable.

For each Actor that should have some type of health bar component (doesn't have to be progress bar), the widget needs to be created on BeginPlay. Look at BP_Postman or BP_BasicEnemyChar for reference. Also make sure that that Actor has Class Settings > Impelemented Interfaces > BPI_Damagable!