All Enemies have attack + bug bug bug
All Enemies now have BPC_DamageSystem and Event Attack that makes sure the damage is applied to the target. The change is still not visible on our player - so that means it might not be having attack on it - but it might also mean that player's health is too big to see changes (Player HP is set to 1000, but the attack is only 10).

Bug#1 = enemies are able to attack each other. Need to disable that, and only make it possible for them to attack player, parcel and wagon.

Bug#2 = Chameleon's tongue is getting stuck in certain positions. Revise the tongue attack.

Bug #3 = Beetleowl Bug hase some buggy behaviour in BT. Sometimes it continuously goes into player regardless of the acceptance range. Might have to look at the Range values and the logic. Currently "IsWithinIdealRange" - so that might be causing problems on attack if enemy is already touching player, so they don't have the need to move.